+++ hooks/src/components/ImageGallery.tsx
@@ -1,31 +1,9 @@
-import React, { useState, useEffect } from 'react';
-
-const API_KEY = import.meta.env.VITE_PIXABAY_API_KEY;
-
-interface Image {
-  id: number;
-  webformatURL: string;
-  tags: string;
-}
+import React, { useState } from 'react';
+import useImageSearch from '../hooks/useImageSearch';
 
 const ImageGallery: React.FC = () => {
-  const [images, setImages] = useState<Image[]>([]);
   const [keyword, setKeyword] = useState('');
-
-  // useEffect를 사용하여 keyword가 변경될 때마다 이미지를 가져오는 로직을 작성해보세요.
-  useEffect(() => {
-    const fetchImages = async () => {
-      try {
-        const response = await fetch(`https://pixabay.com/api?key=${import.meta.env.VITE_PIXABAY_API_KEY}`);
-        const { hits } = await response.json();
-
-        setImages(hits);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
-    fetchImages();
-  }, []);
+  const { images } = useImageSearch(keyword);
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
